Refactor utils and argparse
Separate more functions to `cartridge-cli.utils` module
Move arguments parsing to `cartridge-cli.argparse` module
Rewrite argparse unit test
